Card page refactor applicant add page
