--- frontend/management/src/components/applicant-add/ApplicantAdd.tsx
@@ -1,29 +1,0 @@
-import Button from "../Button.tsx";
-import ApplicantAddCardContainer from "../../containers/applicant-add/ApplicantAddCardContainer.tsx";
-import "../../styles/applicant-add-card.css"
-import {ReactNode} from "react";
-
-function ApplicantAdd({goToApplicantsPage}: Props): ReactNode {
-  return (
-    <div className="applicant-add">
-      <span className="applicant-page__header">
-        <Button
-          handleClick={goToApplicantsPage}
-          btnClasses={"applicant-add__header__back-btn"}
-          iconClass={"bi-arrow-left-short"}
-          spanTextClass={"applicant-add__header__btn__text"}
-          text={"Back to all applicants"}
-        />
-      </span>
-      <div className="applicant-add__container">
-        <ApplicantAddCardContainer/>
-      </div>
-    </div>
-  )
-}
-
-interface Props {
-  goToApplicantsPage: () => void;
-}
-
-export default ApplicantAdd
