+++ frontend/management/src/containers/applicant-add/ApplicantAddPageContainer.tsx
@@ -1,0 +1,20 @@
+import {useNavigate} from "react-router-dom";
+import ApplicantAddPage from "../../components/applicant-add/ApplicantAddPage.tsx";
+import {ReactNode} from "react";
+import CardPageContainer from "../card/CardPageContainer.tsx";
+
+function ApplicantAddPageContainer(): ReactNode {
+  const navigate = useNavigate();
+
+  function goToApplicantsPage(): void {
+    navigate(`/applicants`);
+  }
+
+  return (
+    <CardPageContainer>
+      <ApplicantAddPage goToApplicantsPage={goToApplicantsPage}/>
+    </CardPageContainer>
+  );
+}
+
+export default ApplicantAddPageContainer
